requirements.txt
asttokens==3.0.0
async-timeout==5.0.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
exceptiongroup==1.2.2
executing==2.2.0
Flask==3.1.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
ipykernel==6.29.5
ipython==8.33.0
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pymongo==4.11.2
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-pam==2.0.2
pyzmq==26.2.1
redis==5.2.1
schedule==1.2.2
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
